+++ src/base/b-virtual-scroll/b-virtual-scroll.ts
@@ -256,7 +256,7 @@ export default class bVirtualScroll extends iData implements iItems {
 	 * @override
 	 * @emits chunkLoading(page: number)
 	 */
-	initLoad(data?: unknown, params: InitLoadOptions = {}): CanPromise<void> {
+	initLoad(data?: unknown, opts?: InitLoadOptions): CanPromise<void> {
 		if (!this.lfc.isBeforeCreate()) {
 			this.reInit();
 		}
@@ -265,7 +265,7 @@ export default class bVirtualScroll extends iData implements iItems {
 			this.emit('chunkLoading', 0);
 		}
 
-		return super.initLoad(data, params);
+		return super.initLoad(data, opts);
 	}
 
 	/**
@@ -414,7 +414,7 @@ export default class bVirtualScroll extends iData implements iItems {
 	}
 
 	/** @override */
-	protected syncDataProviderWatcher(): void {
+	protected syncDataProviderWatcher(initLoad?: boolean): void {
 		const
 			provider = this.dataProvider;
 
@@ -422,7 +422,7 @@ export default class bVirtualScroll extends iData implements iItems {
 			this.reInit();
 
 		} else {
-			super.syncDataProviderWatcher();
+			super.syncDataProviderWatcher(initLoad);
 		}
 	}