+++ src/core/component/directives/in-view/adapter.ts
@@ -6,6 +6,8 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+import { deprecate } from 'core/functools/deprecation';
+
 import MutationObserverStrategy from 'core/component/directives/in-view/mutation';
 import IntersectionObserverStrategy from 'core/component/directives/in-view/intersection';
 
@@ -31,7 +33,7 @@ export default class InViewAdapter {
 	protected adaptee?: ObserveStrategy;
 
 	/**
-	 * True if an adapter instance has an adaptee
+	 * True if the adapter instance has an adaptee
 	 */
 	get hasAdaptee(): boolean {
 		return this.adaptee !== undefined;
@@ -46,7 +48,7 @@ export default class InViewAdapter {
 	}
 
 	/**
-	 * Returns true if an adaptee type is 'mutation'
+	 * Returns true if the passed adaptee type is 'mutation'
 	 * @param adaptee
 	 */
 	isMutation(adaptee: ObserveStrategy): adaptee is MutationObserverStrategy {
@@ -54,7 +56,7 @@ export default class InViewAdapter {
 	}
 
 	/**
-	 * Returns true if an adaptee type is 'observer'
+	 * Returns true if the passed adaptee type is 'observer'
 	 * @param adaptee
 	 */
 	isIntersection(adaptee: ObserveStrategy): adaptee is IntersectionObserverStrategy {
@@ -62,12 +64,12 @@ export default class InViewAdapter {
 	}
 
 	/**
-	 * Starts observing the specified element
+	 * Starts to observe the specified element
 	 *
 	 * @param el
 	 * @param params
 	 */
-	observe(el: Element, params: CanArray<InitOptions>): false | void {
+	observe(el: Element, params: CanArray<InitOptions>): false | undefined {
 		if (!this.adaptee) {
 			return false;
 		}
@@ -82,53 +84,85 @@ export default class InViewAdapter {
 	}
 
 	/**
-	 * Activates deactivated elements by the specified group
-	 * @param [group]
+	 * Suspends observing of elements by the specified group.
+	 * Calling this method will temporarily stop observing elements that match the specified group.
+	 * To cancel the suspending invoke `unsuspend`.
+	 *
+	 * @param group
 	 */
-	activate(group?: InViewGroup): void {
-		if (!this.adaptee) {
-			return;
-		}
-
-		return this.adaptee.setGroupState(false, group);
-	}
+	suspend(group: InViewGroup): void;
 
 	/**
-	 * Deactivates elements by the specified group
-	 * @param [group]
-	 */
-	deactivate(group?: InViewGroup): void {
-		if (!this.adaptee) {
-			return;
-		}
-
-		return this.adaptee.setGroupState(true, group);
-	}
-
-	/**
-	 * Removes an element from observable elements
+	 * Suspends observing of the specified element
+	 *
 	 * @param el
+	 * @param threshold - `threshold` should be specified because it's used as a unique key for observables
 	 */
-	remove(el: Element): boolean {
-		if (!this.adaptee) {
-			return false;
-		}
-
-		return this.adaptee.unobserve(el);
+	suspend(el: Element, threshold: number): void;
+	suspend(groupOrElement: InViewGroup | Element, threshold?: number): void {
+		return this.adaptee?.suspend(groupOrElement, threshold);
 	}
 
 	/**
-	 * Stops observing the specified element
+	 * Unsuspends observing of the specified group of elements
+	 * @param group
+	 */
+	unsuspend(group: InViewGroup): void;
+
+	/**
+	 * Unsuspends observing of the specified element
+	 *
+	 * @param el
+	 * @param threshold - `threshold` should be specified because it's used as a unique key for observables
+	 */
+	unsuspend(el: Element, threshold: number): void;
+	unsuspend(groupOrElement: InViewGroup | Element, threshold?: number): void {
+		return this.adaptee?.unsuspend(groupOrElement, threshold);
+	}
+
+	/**
+	 * Re-initializes observing of the specified group
+	 * @param group
+	 */
+	reObserve(group: InViewGroup): void;
+
+	/**
+	 * Re-initializes observing of the specified element
+	 *
+	 * @param el
+	 * @param threshold - `threshold` should be specified because it's used as a unique key for observables
+	 */
+	reObserve(el: Element, threshold: number): void;
+	reObserve(groupOrElement: InViewGroup | Element, threshold?: number): void {
+		return this.adaptee?.reObserve(groupOrElement, threshold);
+	}
+
+	/**
+	 * @see [[InViewAdapter.remove]]
+	 * @deprecated
+	 */
+	stopObserve(el: Element, threshold?: number): boolean {
+		deprecate({
+			type: 'function',
+			alternative: 'inViewAdapter.remove',
+			name: 'inViewAdapter.stopObserve'
+		});
+
+		return this.remove(el, threshold);
+	}
+
+	/**
+	 * Removes the passed element from observable elements
 	 *
 	 * @param el
 	 * @param [threshold]
 	 */
-	stopObserve(el: Element, threshold?: number): boolean {
+	remove(el: Element, threshold?: number): boolean {
 		if (!this.adaptee) {
 			return false;
 		}
 
-		return this.adaptee.stopObserve(el, threshold);
+		return this.adaptee.unobserve(el, threshold);
 	}
 
 	/**