+++ src/core/component/directives/in-view/super.ts
@@ -6,17 +6,21 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import Async from 'core/async';
 import symbolGenerator from 'core/symbol';
+import Async from 'core/async';
 
 import {
 
 	InViewGroup,
 	InitOptions,
+
 	ObservableElement,
 	ObservableElementsThresholdMap,
 	ObservableThresholdMap,
-	Size
+	ObservablesByGroup,
+
+	Size,
+	UnobserveOptions
 
 } from 'core/component/directives/in-view/interface';
 
@@ -30,6 +34,16 @@ export default abstract class AbstractInView {
 	protected readonly elements: ObservableElementsThresholdMap = new Map();
 
 	/**
+	 * Map of observable elements sorted by a group parameter
+	 */
+	protected readonly observablesByGroup: ObservablesByGroup = new Map();
+
+	/**
+	 * Map of elements that was suspended
+	 */
+	protected readonly suspendedElements: ObservableElementsThresholdMap = new Map();
+
+	/**
 	 * Queue of elements that wait to become observable
 	 */
 	protected readonly awaitingElements: ObservableElementsThresholdMap = new Map();
@@ -37,6 +51,7 @@ export default abstract class AbstractInView {
 	/**
 	 * Async instance
 	 */
+	// eslint-disable-next-line @typescript-eslint/no-invalid-this
 	protected readonly async: Async<this> = new Async(this);
 
 	/**
@@ -47,7 +62,7 @@ export default abstract class AbstractInView {
 			POP_AWAITING_INTERVAL = 75;
 
 		this.async.setInterval(() => {
-			if (!this.awaitingElements.size) {
+			if (this.awaitingElements.size === 0) {
 				return;
 			}
 
@@ -61,32 +76,85 @@ export default abstract class AbstractInView {
 	}
 
 	/**
-	 * Activates or deactivates elements by the specified group
+	 * Suspends the specified element or elements by the specified group
 	 *
-	 * @param isDeactivated
-	 * @param [group]
+	 * @param groupOrElement
+	 * @param [threshold]
 	 */
-	setGroupState(isDeactivated: boolean, group?: InViewGroup): void {
-		this.maps().forEach((map) => {
-			map.forEach((el) => {
-				if (!group || el.group !== group) {
-					return;
-				}
+	suspend(groupOrElement: InViewGroup | Element, threshold?: number): void {
+		if (groupOrElement instanceof Element) {
+			const
+				{suspendedElements} = this;
 
-				el.isDeactivated = isDeactivated;
+			if (threshold == null) {
+				return;
+			}
+
+			const
+				observable = this.getEl(groupOrElement, threshold);
+
+			if (observable == null) {
+				return;
+			}
+
+			let
+				map = suspendedElements.get(groupOrElement);
+
+			if (map == null) {
+				suspendedElements.set(groupOrElement, map = new Map());
+			}
+
+			map.set(threshold, observable);
+
+		} else {
+			this.observablesByGroup.get(groupOrElement)?.forEach((observable) => {
+				this.unobserve(observable.node, {threshold: observable.threshold, suspend: true});
 			});
-		});
+		}
 	}
 
 	/**
-	 * Returns an observable element parameters
+	 * Unsuspends the specified element or elements by the specified group
+	 *
+	 * @param groupOrElement
+	 * @param [threshold]
+	 */
+	unsuspend(groupOrElement: InViewGroup | Element, threshold?: number): void {
+		if (groupOrElement instanceof Element) {
+			const
+				{suspendedElements} = this;
+
+			if (threshold == null) {
+				return;
+			}
+
+			const
+				map = suspendedElements.get(groupOrElement),
+				observable = map?.get(threshold);
+
+			if (observable == null || map == null) {
+				return;
+			}
+
+			map.delete(threshold);
+			this.observe(observable.node, observable);
+
+		} else {
+			this.observablesByGroup.get(groupOrElement)?.forEach((observable) => {
+				this.observe(observable.node, observable);
+			});
+		}
+	}
+
+	/**
+	 * Returns an observable element
 	 *
 	 * @param el
 	 * @param threshold
 	 */
 	getEl(el: Element, threshold: number): CanUndef<ObservableElement> {
 		const map = this.getThresholdMap(el);
-		return map && map.get(threshold);
+		return map?.get(threshold);
 	}
 
 	/**
@@ -102,10 +170,6 @@ export default abstract class AbstractInView {
 	 * @param observable
 	 */
 	call(observable: ObservableElement): void {
-		if (observable.isDeactivated) {
-			return;
-		}
-
 		const count = Object.isFunction(observable.count) ?
 			observable.count() :
 			observable.count;
@@ -119,7 +183,7 @@ export default abstract class AbstractInView {
 		}
 
 		if (observable.once) {
-			this.stopObserve(observable.node, observable.threshold);
+			this.unobserve(observable.node, observable.threshold);
 		}
 	}
 
@@ -135,7 +199,21 @@ export default abstract class AbstractInView {
 		}
 
 		const
+			{observablesByGroup} = this,
 			observable = this.createObservable(el, opts);
+
+		if (observable.group != null) {
+			let observableSet: Set<ObservableElement>;
+
+			if (observablesByGroup.has(observable.group)) {
+				observableSet = observablesByGroup.get(observable.group)!;
+
+			} else {
+				observablesByGroup.set(observable.group, observableSet = new Set());
+			}
+
+			observableSet.add(observable);
+		}
 
 		if (observable.wait) {
 			this.putInMap(this.awaitingElements, observable);
@@ -148,62 +226,48 @@ export default abstract class AbstractInView {
 	}
 
 	/**
-	 * Stops to observe the specified element
+	 * Re-initializes observation of the specified element or group
 	 *
-	 * @param el
+	 * @param groupOrElement
 	 * @param threshold
 	 */
-	stopObserve(el: Element, threshold?: number): boolean {
-		const
-			thresholdMap = this.getThresholdMap(el);
+	reObserve(groupOrElement: InViewGroup | Element, threshold?: number): void {
+		if (threshold != null) {
+			this.suspend(<Element>groupOrElement, threshold);
+			this.unsuspend(<Element>groupOrElement, threshold);
 
-		const stopObserve = (o) => {
-			this.clearAllAsync(o);
-
-			if (o.removeStrategy === 'remove') {
-				return this.unobserve(el, threshold);
-			}
-
-			o.isDeactivated = true;
-			return true;
-		};
-
-		if (thresholdMap && threshold === undefined) {
-			thresholdMap.forEach((observable) => {
-				stopObserve(observable);
-			});
-
-			return true;
+		} else {
+			this.suspend(<InViewGroup>groupOrElement);
+			this.unsuspend(<InViewGroup>groupOrElement);
 		}
-
-		if (thresholdMap && threshold !== undefined) {
-			const
-				observable = thresholdMap.get(threshold);
-
-			if (!observable) {
-				return false;
-			}
-
-			return stopObserve(observable);
-		}
-
-		return false;
 	}
 
 	/**
-	 * Removes an element from observable elements
+	 * Removes or suspends the specified element
 	 *
 	 * @param el
-	 * @param [threshold]
+	 * @param unobserveOptionsOrThreshold
 	 */
-	unobserve(el: Element, threshold?: number): boolean {
+	unobserve(el: Element, unobserveOptionsOrThreshold?: UnobserveOptions | number): boolean {
+		let
+			threshold: CanUndef<number>,
+			suspend: CanUndef<boolean>;
+
+		if (Object.isNumber(unobserveOptionsOrThreshold)) {
+			threshold = unobserveOptionsOrThreshold;
+
+		} else if (Object.isPlainObject(unobserveOptionsOrThreshold)) {
+			threshold = unobserveOptionsOrThreshold.threshold;
+			suspend = unobserveOptionsOrThreshold.suspend;
+		}
+
 		const
 			map = this.getElMap(el),
 			thresholdMap = this.getThresholdMap(el);
 
 		if (thresholdMap && threshold === undefined) {
 			thresholdMap.forEach((observable) => {
-				this.remove(observable);
+				this.remove(observable, suspend);
 			});
 
 			return map.delete(el);
@@ -217,7 +281,7 @@ export default abstract class AbstractInView {
 				return false;
 			}
 
-			this.remove(observable);
+			this.remove(observable, suspend);
 			return thresholdMap.delete(threshold);
 		}
 
@@ -239,10 +303,8 @@ export default abstract class AbstractInView {
 			count: true,
 
 			isLeaving: false,
-			isDeactivated: false,
-			removeStrategy: 'remove',
 
-			// @ts-ignore
+			// @ts-expect-error (override)
 			threshold: 1,
 
 			size: {
@@ -263,7 +325,7 @@ export default abstract class AbstractInView {
 	}
 
 	/**
-	 * Puts an observable element in the specified map
+	 * Puts an observable element into the specified map
 	 *
 	 * @param map
 	 * @param observable
@@ -287,7 +349,7 @@ export default abstract class AbstractInView {
 	}
 
 	/**
-	 * All maps combined into one
+	 * Returns all maps combined into one
 	 */
 	protected maps(): ObservableElementsThresholdMap {
 		return new Map([
@@ -297,7 +359,7 @@ export default abstract class AbstractInView {
 	}
 
 	/**
-	 * Returns a map which contains an element
+	 * Returns a map which contains the specified element
 	 * @param el
 	 */
 	protected getElMap(el: Element): ObservableElementsThresholdMap {
@@ -305,7 +367,7 @@ export default abstract class AbstractInView {
 	}
 
 	/**
-	 * Initializes element from an observable queue
+	 * Initializes an element from the observable queue
 	 */
 	protected popAwaiting(): void {
 		const
@@ -314,7 +376,7 @@ export default abstract class AbstractInView {
 		awaitingElements.forEach((map, node) => {
 			map.forEach((observable, threshold) => {
 				const
-					isResolved = Boolean(observable.wait && observable.wait());
+					isResolved = Boolean(observable.wait?.());
 
 				if (!isResolved) {
 					return;
@@ -333,6 +395,9 @@ export default abstract class AbstractInView {
 
 	/**
 	 * Sets a size of the specified observable element
+	 *
+	 * @param observable
+	 * @param size
 	 */
 	protected setObservableSize(observable: ObservableElement, size: Size): void {
 		observable.size.width = size.width;
@@ -357,15 +422,37 @@ export default abstract class AbstractInView {
 	 * Initializes observing for the specified element
 	 * @param observable
 	 */
+	// eslint-disable-next-line @typescript-eslint/no-unused-vars-experimental
 	protected initObserve(observable: ObservableElement): CanUndef<ObservableElement> {
 		return undefined;
 	}
 
 	/**
-	 * Removes element from observer data
+	 * Removes an element from the observer data
+	 *
 	 * @param observable
+	 * @param [suspend]
 	 */
-	protected remove(observable: ObservableElement): boolean {
-		return false;
+	// eslint-disable-next-line @typescript-eslint/no-unused-vars-experimental
+	protected remove(observable: ObservableElement, suspend?: boolean): boolean {
+		if (!suspend) {
+			if (observable.group != null) {
+				this.observablesByGroup.get(observable.group)?.delete(observable);
+			}
+
+			const
+				map = this.suspendedElements.get(observable.node);
+
+			if (map) {
+				map.delete(observable.threshold);
+
+				if (map.size === 0) {
+					this.suspendedElements.delete(observable.node);
+				}
+			}
+		}
+
+		this.clearAllAsync(observable);
+		return true;
 	}
 }