+++ src/super/i-block/i-block.ts
@@ -17,10 +17,10 @@ import symbolGenerator from 'core/symbol';
 import { deprecated } from 'core/functools';
 
 import SyncPromise from 'core/promise/sync';
-import Async, { AsyncOptions, ClearOptionsId, ProxyCb, BoundFn } from 'core/async';
 import log, { LogMessageOptions } from 'core/log';
 
 import { EventEmitter2 as EventEmitter } from 'eventemitter2';
+import Async, { AsyncOptions, ClearOptionsId, ProxyCb, BoundFn } from 'core/async';
 
 //#if runtime has core/helpers
 import * as helpers from 'core/helpers';
@@ -1828,7 +1828,8 @@ export default abstract class iBlock extends ComponentInterface {
 	 *
 	 * @param [data] - data object (for events)
 	 * @param [opts] - additional options
-	 * @emits `initLoad(data: CanUndef<unknown>, params: CanUndef<InitLoadOptions>)`
+	 * @emits `initLoadStart(options: CanUndef<InitLoadOptions>)`
+	 * @emits `initLoad(data: CanUndef<unknown>, options: CanUndef<InitLoadOptions>)`
 	 */
 	@hook('beforeDataCreate')
 	initLoad(data?: unknown | InitLoadCb, opts: InitLoadOptions = {}): CanPromise<void> {
@@ -1837,6 +1838,10 @@ export default abstract class iBlock extends ComponentInterface {
 		}
 
 		this.beforeReadyListeners = 0;
+
+		if (opts.emitStartEvent !== false) {
+			this.emit('initLoadStart', opts);
+		}
 
 		if (!opts.silent) {
 			this.componentStatus = 'loading';