+++ src/super/i-page/i-page.ts
@@ -59,7 +59,7 @@ export default abstract class iPage extends iData implements iVisible {
 	 */
 	set pageTitle(value: string) {
 		if (this.isActivated) {
-			this.r.setPageTitle(value, this);
+			void this.r.setPageTitle(value, this);
 		}
 	}
 
@@ -68,7 +68,7 @@ export default abstract class iPage extends iData implements iVisible {
 	 * @see [[iPage.scrollTo]]
 	 */
 	get scrollToProxy(): this['scrollTo'] {
-		return this.async.proxy(this.scrollTo, {
+		return this.async.proxy(this.scrollTo.bind(this), {
 			single: false,
 			label: $$.scrollTo
 		});
@@ -104,10 +104,10 @@ export default abstract class iPage extends iData implements iVisible {
 
 		const scroll = (opts: ScrollToOptions) => {
 			try {
-				window.scrollTo(opts);
+				globalThis.scrollTo(opts);
 
 			} catch {
-				window.scrollTo(opts.left == null ? pageXOffset : opts.left, opts.top == null ? pageYOffset : opts.top);
+				globalThis.scrollTo(opts.left == null ? pageXOffset : opts.left, opts.top == null ? pageYOffset : opts.top);
 			}
 		};
 
@@ -136,22 +136,23 @@ export default abstract class iPage extends iData implements iVisible {
 	 */
 	@watch('!:onStageChange')
 	protected syncStageTitles(): CanUndef<string> {
-		if (!this.stagePageTitles) {
+		const
+			stageTitles = this.stagePageTitles;
+
+		if (stageTitles == null) {
 			return;
 		}
 
-		const
-			stageTitles = this.stage != null && this.stagePageTitles;
-
-		if (stageTitles) {
+		if (this.stage != null) {
 			let
-				v = stageTitles[<string>this.stage];
+				v = stageTitles[this.stage];
 
-			if (!v) {
+			if (v == null) {
 				v = stageTitles['[[DEFAULT]]'];
 			}
 
-			if (v) {
+			// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+			if (v != null) {
 				return this.pageTitle = this.t(Object.isFunction(v) ? v(this) : v);
 			}
 		}
@@ -162,7 +163,7 @@ export default abstract class iPage extends iData implements iVisible {
 	 */
 	@hook('created')
 	protected initTitle(): void {
-		if (!this.syncStageTitles() && this.pageTitleStore) {
+		if (this.syncStageTitles() == null && Object.isTruly(this.pageTitleStore)) {
 			this.pageTitle = this.pageTitleStore;
 		}
 	}