+++ tests/cases.js
@@ -25,5 +25,8 @@ module.exports = [
 	'--name b-button --suit demo',
 
 	// v-image
-	'--name b-dummy --test-entry core/component/directives/image/test'
+	'--name b-dummy --test-entry core/component/directives/image/test',
+
+	// in-view
+	'--name b-dummy --test-entry core/component/directives/in-view/test'
 ];