+++ src/core/component/directives/in-view/CHANGELOG.md
@@ -9,6 +9,24 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.49 (2020-08-03)
+
+#### :boom: Breaking Change
+
+* Removed `isDeactivated`, `removeStrategy` from `observableElement`
+
+#### :rocket: New Feature
+
+* Added `suspend`, `unsuspend`, `reObserve` methods
+
+#### :bug: Bug Fix
+
+* Fixed an issue with `polling` strategy won't fire a `callback`
+
+#### :house: Internal
+
+* Fixed ESLint warnings
+
 ## v3.0.0-rc.19 (2020-05-26)
 
 #### :rocket: New Feature