+++ src/super/i-block/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.48 (2020-08-02)
+
+#### :rocket: New Feature
+
+* Added `initLoadStart` event
+
 ## v3.0.0-rc.39 (2020-07-23)
 
 #### :house: Internal