+++ src/traits/i-history/CHANGELOG.md
@@ -9,6 +9,16 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.49 (2020-08-03)
+
+#### :bug: Bug Fix
+
+* Fixed an issue with `trigger` in `b-bottom-slide` won't fire `v-in-view` a callback (`mutation strategy`)
+
+#### :house: Internal
+
+* Fixed ESLint warnings
+
 ## v3.0.0-rc.16 (2020-05-21)
 
 #### :bug: Bug Fix